Revert "merge changes from master"
This reverts commit 077fc421108c5cc5590fba061e2e6ea6d3e13a6c.

+++ packages/cli/src/evaluation.ee/test-runs.controller.ee.ts
@@ -1,13 +1,7 @@
-import express from 'express';
-import { InstanceSettings } from 'n8n-core';
-
 import { TestRunRepository } from '@/databases/repositories/test-run.repository.ee';
-import { Delete, Get, Post, RestController } from '@/decorators';
-import { ConflictError } from '@/errors/response-errors/conflict.error';
+import { Delete, Get, RestController } from '@/decorators';
 import { NotFoundError } from '@/errors/response-errors/not-found.error';
-import { NotImplementedError } from '@/errors/response-errors/not-implemented.error';
 import { TestRunsRequest } from '@/evaluation.ee/test-definitions.types.ee';
-import { TestRunnerService } from '@/evaluation.ee/test-runner/test-runner.service.ee';
 import { listQueryMiddleware } from '@/middlewares';
 import { getSharedWorkflowIds } from '@/public-api/v1/handlers/workflows/workflows.service';
 
@@ -18,12 +12,9 @@ export class TestRunsController {
 	constructor(
 		private readonly testDefinitionService: TestDefinitionService,
 		private readonly testRunRepository: TestRunRepository,
-		private readonly testRunnerService: TestRunnerService,
-		private readonly instanceSettings: InstanceSettings,
 	) {}
 
-	/**
-	 * This method is used in multiple places in the controller to get the test definition
+	/** This method is used in multiple places in the controller to get the test definition
 	 * (or just check that it exists and the user has access to it).
 	 */
 	private async getTestDefinition(
@@ -43,23 +34,6 @@ export class TestRunsController {
 		return testDefinition;
 	}
 
-	/**
-	 * Get the test run (or just check that it exists and the user has access to it)
-	 */
-	private async getTestRun(
-		req: TestRunsRequest.GetOne | TestRunsRequest.Delete | TestRunsRequest.Cancel,
-	) {
-		const { id: testRunId, testDefinitionId } = req.params;
-
-		const testRun = await this.testRunRepository.findOne({
-			where: { id: testRunId, testDefinition: { id: testDefinitionId } },
-		});
-
-		if (!testRun) throw new NotFoundError('Test run not found');
-
-		return testRun;
-	}
-
 	@Get('/:testDefinitionId/runs', { middlewares: listQueryMiddleware })
 	async getMany(req: TestRunsRequest.GetMany) {
 		const { testDefinitionId } = req.params;
@@ -71,43 +45,33 @@ export class TestRunsController {
 
 	@Get('/:testDefinitionId/runs/:id')
 	async getOne(req: TestRunsRequest.GetOne) {
+		const { id: testRunId, testDefinitionId } = req.params;
+
 		await this.getTestDefinition(req);
 
-		return await this.getTestRun(req);
+		const testRun = await this.testRunRepository.findOne({
+			where: { id: testRunId, testDefinition: { id: testDefinitionId } },
+		});
+
+		if (!testRun) throw new NotFoundError('Test run not found');
+
+		return testRun;
 	}
 
 	@Delete('/:testDefinitionId/runs/:id')
 	async delete(req: TestRunsRequest.Delete) {
-		const { id: testRunId } = req.params;
+		const { id: testRunId, testDefinitionId } = req.params;
 
-		// Check test definition and test run exist
 		await this.getTestDefinition(req);
-		await this.getTestRun(req);
+
+		const testRun = await this.testRunRepository.findOne({
+			where: { id: testRunId, testDefinition: { id: testDefinitionId } },
+		});
+
+		if (!testRun) throw new NotFoundError('Test run not found');
 
 		await this.testRunRepository.delete({ id: testRunId });
 
 		return { success: true };
-	}
-
-	@Post('/:testDefinitionId/runs/:id/cancel')
-	async cancel(req: TestRunsRequest.Cancel, res: express.Response) {
-		if (this.instanceSettings.isMultiMain) {
-			throw new NotImplementedError('Cancelling test runs is not yet supported in multi-main mode');
-		}
-
-		const { id: testRunId } = req.params;
-
-		// Check test definition and test run exist
-		await this.getTestDefinition(req);
-		const testRun = await this.getTestRun(req);
-
-		if (this.testRunnerService.canBeCancelled(testRun)) {
-			const message = `The test run "${testRunId}" cannot be cancelled`;
-			throw new ConflictError(message);
-		}
-
-		await this.testRunnerService.cancelTestRun(testRunId);
-
-		res.status(202).json({ success: true });
 	}
 }
